Remove div/mod by zero errors
Allowing div/mod by zero to return inf/nan has some plausible use cases.
Without altering anything else in the VM, e.g. you can capture the
result of a div by zero, use it as an array index and access it with the
string "inf."

  a[1/0] = 5
  a["inf"]      // Prints 5

As long as this has some usefulness, it's more in-line with the spirit
of the language to allow this behavior, instead of throwing an error.