Array improvements
Improves (implements?) the logic behind array vs hash insertion/lookup.
This also adds proprietary str -> int conversion for array indexing. In
the VM, strings are liberally coerced to numbers based on the operation,
but for indexing arrays such that a["0"] == a[0] but not something like
a["0a"], we need to make sure the entire string is a valid number.
Otherwise, treat it like a unique string.