Parse functions with their own parser
Unfortunately, this effectively sandboxes all user functions, rendering
all outside locals invisible to a given function. The only other
possibility I see at the moment is to implement full lexical scoping
(like Lua). However, that requires all user functions to be defined (or
just declared?) before they're called, which I don't particularly care
for.

I also don't see this as the worst solution since if you need a variable
to be accessible by a function, you can use a global or pass it as an
argument.