Add support for anonymous functions
This parses and compiles anonymous functions inside expressions. Anon
functions are added to the code object's constants pool and utilized as
needed at runtime. Unlike named functions which are added to the global
rf_env struct and added to the global variables upon VM initalization.
This basically treats anonymous functions as a type of constant or
literal value.