Implement generic "foreach" style for loops
The following syntax:

  for x in y {...}
  for x,y in z {...}

Now initializes a loop structure in the VM and iterates over each value
or key/value pair in a given value. If the value is a number, it simply
loops from 0 to the number (inclusive). So this loop:

  for i in 100 {...}

will assign the numbers 0-100 to local variable `i` upon each iteration.
The variable `i` is local to the loop and will not be visible to any
control outside of the loop (maybe temporary). Currently, floats are
truncated to integers.