Implement lexer modes
This utilizes a parameter in the lexer struct to determine whether the
parse expects a nud or led token. This is necessary to implement regex
literals in the future, so the lexer can pre-compile the regex literals
instead of returning a single `/` token. An added benefit of the lexer
mode is this eliminates extraneous OP_NUM and OP_NEG instructions for
numeric literals with the prefixed `+` or `-`, though this only works if
the unary operator is directly prefixed to the number without any
whitespace.

This commit also separates common utility functions to util.c, which
allows me to unify certain things like numeric string parsing, which is
performed at various stages throughout the interpreter. Currently the
u_str2d and u_str2i64 functions serve as wrappers for strtod and
strtoll/stroull, respectively.

Finally, this commit introduces support for arbitrary underscores in
numeric literals.

Squashed commit of the following:

commit 03cd4c6e314229341e8b73ef1bdcfb8ad65a199f
Author: Darryl Abbate <rootbeersoup@gmail.com>
Date:   Sat Jan 30 17:05:48 2021 -0600

    Replace all strtod/strtoll calls

commit a386b64ab3e274b675226ee2f93c80a7155470c8
Author: Darryl Abbate <rootbeersoup@gmail.com>
Date:   Sat Jan 30 01:21:16 2021 -0600

    Commit working version of u_str2d()

    Still want to clean the util functions up and replace all
    strtoll/strtoull/strtod calls throughout the interpreter with
    u_str2i64/u_str2d.

commit bd1508bb5b05e953f58dd61f0a729dde76bddac0
Author: Darryl Abbate <rootbeersoup@gmail.com>
Date:   Thu Jan 28 20:18:50 2021 -0600

    More utils work

commit f93894c500e85c4562ba60b534a678a0da78eb9e
Author: Darryl Abbate <rootbeersoup@gmail.com>
Date:   Tue Jan 26 16:38:38 2021 -0600

    Begin implementing utils

    This also adds experiemtnal support for underscores in numeric literals
    (currently integers only)

commit 238d7ec1988e44de8a4beb5ea8faf20ab61f3311
Author: Darryl Abbate <rootbeersoup@gmail.com>
Date:   Sat Jan 23 23:49:48 2021 -0600

    Begin implementing lexer mode