Use PCRE2 instead of POSIX regex.h
This replaces the system regex.h with PCRE2. For now, this requires
PCRE2 to be installed somewhere on the system where pkg-config can find
it. I'll probably include the PCRE2 source code in /lib as a Git
submodule in case PCRE2 can't be found on a user's system.

New library functions: sub(), gsub()

Caveats so far:
- PCRE2 interprets `\nnn` as octal, unless it's a backreference. This is
  in contrast to how Riff parses `\nnn` as decimal in string and
  character literals.

Still to-do before merging:
- Refactor split() to use regexes as the optional delimiter
- Compile-time (lexer) error handling for regex literals
- Repurpose `$` operator to extract capture group substrings
- Utilize PCRE2's JIT functionality (maybe)