Interpret `\nnn` as octal
This change is simply to match PCRE2, which interprets `\nnn` as octal
(unless it's interpreted as a backreference). Some quick research shows
that interpreting `\nnn` as decimal is much less common than octal. Lua
was the only remotely popular language I could find that interprets
decimal. Some other will interpret as hex, but the majority interpret as
octal.