Repurpose `$` operator for capture access
This changes the `$` operator to read/write to a table that stores
capured regex subexpressions instead of command-line arguments. For now,
every match operation stores all subexpressions into numbered indices
starting from 0, where 0 is the matched substring and each succeeding
number corresponds to any captured subexpressions.

A potential optimization would be to store captured subexpressions only
when the field vector is accessed. The third option was to store a
global pcre2_match_data pointer in the VM and only allow read access to
the field vector. This would throw away a lot of the work done to make
`$n` expressions work properly, and would arguably be a less useful
feature.

Note that the `arg` table still exists for accessing command-line
arguments.