Store captured substrings from `gsub()`/`sub()`
This is kind of a dirty solution, but this works by registering and
storing the VM's fields table with the regex file. A more principled
approach would be to store the fields table in the rf_env and passing
its pointer around.

Note that this still only stores numbered captures, not named captures.