Don't consume `-` prefixing numeric literals
This was an oversight on my part. Consuming the leading unary minus sign
and simply lexing the string as a negative number interferes with
operator precedence when succeeding operators have higher precedence
than negation.

This now causes the "INT64_MIN" test to fail since it will now be parsed
as a floating-point number. This is unavoidable without changing how
compile-time constants are stored and retrieved.

Note that this is still acceptable for leading unary plus signs; the
result of the expression is unaffected either way since the unary plus
is superfluous.