Add support for multicharacter literals
This is obviously a niche feature, but it doesn't directly interfere
with typical character literals. These work similarly to multicharacter
constants in languages like C, where the character sequence is
interpreted as a big-endian zero-padded integer composed of the ASCII
codes concatenated together.

For now, in the event of overflow, the resulting integer will be the
lowest 64 bits of the sequence. It may or may not be worth deferring to
doubles in the event of overflow. Regardless, I don't think it's
necessary or particularly useful to simply throw an error if a
multicharacter literal overflows.