Add extra fields to riff_str
This adds a 'hints' field for string coercion hints and an 'extra' field
to contain lexical token data for reserved words. Storing the
corresponding token value in the riff_str struct allows the lexer to
easily determine if a word is a unique identifier or keyword without the
ugly switch/case state machine.