Update WASM config
Notably, this exposes a new function `riff_version()` in the WASM module
for the JS to fetch the version string without it being hardcoded. An
alternative to exposing a C function would be to simply append the
version string to the riff.js file spit out by Emscripten, e.g.

  const riffVersion = "$(RIFF_VERSION)";